# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Reserva de turnos odontológicos",
        "title": "Api de turnos Odontologicos",
        "termsOfService": "https://developers.ctd.com.ar/es_ar/terminos-y-condiciones",
        "contact": {
            "name": "API Support"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "paths": {
...
        "/turnos/{idTurno}": {
            "get": {
                "description": "Retorna un turno dado su ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Turnos"
                ],
                "summary": "Obtener turno por ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID del turno a obtener",
                        "name": "idTurno",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Turno obtenido",
                        "schema": {
                            "$ref": "#/definitions/domain.Turno"
                        }
                    },
                    "400": {
                        "description": "Error de solicitud",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "No se encontró el turno",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorResponse"
                        }
                    }
                }
            },
            "put": {
                "description": "Actualiza un turno existente por su ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Turnos"
                ],
                "summary": "Actualizar turno",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID del turno a actualizar",
                        "name": "idTurno",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Información actualizada del turno",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/domain.Turno"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Turno actualizado",
                        "schema": {
                            "$ref": "#/definitions/domain.Turno"
                        }
                    },
                    "400": {
                        "description": "Error de solicitud",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorResponse"
                        }
                    }
                }
            },
            "post": {
                "description": "Retorna un turno dado su ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Turnos"
                ],
                "summary": "Obtener turno por ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID del turno a obtener",
                        "name": "idTurno",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Turno obtenido",
                        "schema": {
                            "$ref": "#/definitions/domain.Turno"
                        }
                    },
                    "400": {
                        "description": "Error de solicitud",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorResponse"
                        }
                    }
                }
            },
            "delete": {
                "description": "Elimina un turno por su ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Turnos"
                ],
                "summary": "Eliminar un turno",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID del turno a eliminar",
                        "name": "idTurno",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "El turno ha sido correctamente eliminado",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Error de solicitud",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorResponse"
                        }
                    }
                }
            },
            "patch": {
                "description": "Actualiza campos específicos de un turno por su ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Turnos"
                ],
                "summary": "Actualizar campos específicos de un turno",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID del turno a actualizar",
                        "name": "idTurno",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Campos a actualizar del turno",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handler.Request"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Turno actualizado",
                        "schema": {
                            "$ref": "#/definitions/domain.Turno"
                        }
                    },
                    "400": {
                        "description": "Error de solicitud",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
...
        "domain.Turno": {
            "type": "object",
            "required": [
                "descripcionTurno",
                "fechaTurno",
                "idOdontologo",
                "idPaciente"
            ],
            "properties": {
                "descripcionTurno": {
                    "type": "string"
                },
                "fechaTurno": {
                    "type": "string"
                },
                "idOdontologo": {
                    "type": "string"
                },
                "idPaciente": {
                    "type": "string"
                },
                "idTurno": {
                    "type": "integer"
                }
            }
        },
        "handler.ErrorResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "message": {
                    "type": "string"
                }
            }
        },
        "handler.Request": {
            "type": "object",
            "properties": {
                "apellidoOdontologo": {
                    "type": "string"
                },
                "matriculaOdontologo": {
                    "type": "string"
                },
                "nombreOdontologo": {
                    "type": "string"
                }
            }
        }
    }
}